fix: 게임 시작 시 오디오 연결 끊김 수정
- 게임 시작 시 오디오 연결 끊기는 오류 수정
- 오류 수정으로 인한 페이지 구조 변경
