+++ packages/frontend/src/App.tsx
@@ -7,8 +7,6 @@ import { ThemeProvider } from "styled-components";
 
 import GlobalStyles from "./global-styles/global-styles";
 import { theme } from "./global-styles/theme";
-import CanvasPage from "./pages/CanvasPage/CanvasPage.component";
-import GamePage from "./pages/GamePage/GamePage.component";
 import LandingPage from "./pages/LandingPage/LandingPage.component";
 import RoomPage from "./pages/RoomPage/RoomPage.component";
 import { ratioAtom } from "./store/ratio";
@@ -47,9 +45,7 @@ const App = () => {
       <AtomsDevtools>
         <Routes>
           <Route path="/" element={<LandingPage />}></Route>
-          <Route path="canvas" element={<CanvasPage />}></Route>
-          <Route path="room" element={<RoomPage />}></Route>
-          <Route path="game" element={<GamePage />}></Route>
+          <Route path="room/*" element={<RoomPage />}></Route>
         </Routes>
       </AtomsDevtools>
       <GlobalStyles />
