+++ packages/frontend/src/components/Game/Gartic/Gartic.styles.tsx
